refactor: 패키지를 정리한다
