+++ src/main/java/study/racinggame/domain/engine/MovingStrategy.java
@@ -1,4 +1,4 @@
-package study.racinggame.domain.car;
+package study.racinggame.domain.engine;
 
 public interface MovingStrategy {
     boolean movable();
