+++ src/test/java/study/racinggame/domain/engine/EngineTest.java
@@ -1,8 +1,7 @@
-package study.racinggame.domain;
+package study.racinggame.domain.engine;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import study.racinggame.domain.car.Engine;
 
 import static org.assertj.core.api.Assertions.assertThat;
